# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  activity.ActivityItem:
    properties:
      album_art_url:
        type: string
      album_name:
        type: string
      artist_name:
        type: string
      avatar_id:
        type: string
      id:
        type: string
      is_playing:
        type: boolean
      platform:
        type: string
      played_at:
        type: string
      preview_url:
        type: string
      spotify_url:
        type: string
      started_at:
        type: string
      track_id:
        type: string
      track_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  auth.UserProfile:
    properties:
      avatar_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_premium:
        type: boolean
      user_tag:
        type: string
      username:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: API documentation for the Chen, a music compatibility and social platform.
  title: Chen API
  version: "1.0"
paths:
  /activity/feed:
    get:
      description: Fetch the user's personalized activity feed from friends and followed
        users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/activity.ActivityItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Activity Feed
      tags:
      - activity
...
  /auth/signout:
    post:
      description: Sign out the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Sign Out
      tags:
      - auth
  /auth/user:
    get:
      description: Fetch the current authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserProfile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Current User
      tags:
      - auth
...
swagger: "2.0"